Day 4-  Build REST Controller

+++ backend/src/main/java/com/internship/tool/service/ComplianceService.java
@@ -2,7 +2,11 @@ package com.internship.tool.service;
 
 import com.internship.tool.entity.Compliance;
 import com.internship.tool.repository.ComplianceRepository;
+import com.internship.tool.exception.ResourceNotFoundException;
+
 import org.springframework.stereotype.Service;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
@@ -15,23 +19,29 @@ public class ComplianceService {
         this.repository = repository;
     }
 
-    // Create
+    
     public Compliance createCompliance(Compliance compliance) {
         return repository.save(compliance);
     }
 
-    // Get all
+    
     public List<Compliance> getAllCompliance() {
         return repository.findAll();
     }
 
-    // Get by ID
-    public Compliance getComplianceById(Long id) {
-        return repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Compliance not found with id: " + id));
+    
+    public Page<Compliance> getAllPaginated(Pageable pageable) {
+        return repository.findAll(pageable);
     }
 
-    // Update
+   
+    public Compliance getComplianceById(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(
+                        "Compliance not found with id: " + id));
+    }
+
+    
     public Compliance updateCompliance(Long id, Compliance updated) {
         Compliance existing = getComplianceById(id);
 
@@ -45,8 +55,12 @@ public class ComplianceService {
         return repository.save(existing);
     }
 
-    // Delete
+    // ✅ DELETE
     public void deleteCompliance(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(
+                    "Compliance not found with id: " + id);
+        }
         repository.deleteById(id);
     }
 }
